requirements.txt
dashscope==1.25.5
streamlit==1.52.2
pandas==2.3.3
torch==2.2.2
faiss-cpu==1.12.0
pillow==12.0.0
openai==2.14.0
transformers==4.57.3
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-core==1.2.5
numpy==1.26.4
pypdf==6.5.0
safetensors==0.7.0
